check for initial installation

+++ etc/insights-client.spec
@@ -57,40 +57,43 @@ getent passwd insights > /dev/null || \
     -c "Red Hat Insights" -d /var/lib/insights
 
 %post
-#Migrate existing machine-id
-if  [ -f "/etc/redhat_access_proactive/machine-id" ]; then
-mv /etc/redhat_access_proactive/machine-id /etc/insights-client/machine-id
+# Only perform migration from redhat-access-insights to insights-client
+if  [ $1 -eq 1  ]; then
+	#Migrate existing machine-id
+	if  [ -f "/etc/redhat_access_proactive/machine-id" ]; then
+		mv /etc/redhat_access_proactive/machine-id /etc/insights-client/machine-id
+	fi
+	#Migrate OTHER existing machine-id
+	if [ -f "/etc/redhat-access-insights/machine-id" ]; then
+		mv /etc/redhat-access-insights/machine-id /etc/insights-client/machine-id
+	fi
+	#Migrate existing config
+	if [ -f "/etc/redhat-access-insights/redhat-access-insights.conf" ]; then
+		mv /etc/redhat-access-insights/redhat-access-insights.conf /etc/insights-client/insights-client.conf
+		sed -i 's/\[redhat-access-insights\]/\[insights-client\]/' /etc/insights-client/insights-client.conf
+	fi
+	#Migrate registration record
+	if [ -f "/etc/redhat-access-insights/.registered" ]; then
+		mv /etc/redhat-access-insights/.registered /etc/insights-client/.registered
+	fi
+	#Migrate last upload record
+	if [ -f "/etc/redhat-access-insights/.lastupload" ]; then
+		mv /etc/redhat-access-insights/.lastupload /etc/insights-client/.lastupload
+	fi
+	if ! [ -d "/etc/redhat-access-insights" ]; then
+		mkdir /etc/redhat-access-insights
+	fi
+	# Symlink new cron job if the old one exists. Remove the old one
+	if [ -f "/etc/cron.daily/redhat-access-insights" ]; then
+		rm -f /etc/cron.daily/redhat-access-insights
+		ln -sf /etc/insights-client/insights-client.cron /etc/cron.daily/insights-client                               
+	fi 
+	ln -sf /etc/insights-client/insights-client.conf /etc/redhat-access-insights/redhat-access-insights.conf
+	ln -sf /etc/insights-client/insights-client.cron /etc/redhat-access-insights/redhat-access-insights.cron
+	ln -sf /etc/insights-client/.registered /etc/redhat-access-insights/.registered
+	ln -sf /etc/insights-client/.unregistered /etc/redhat-access-insights/.unregistered
+	ln -sf /etc/insights-client/machine-id /etc/redhat-access-insights/machine-id
 fi
-#Migrate OTHER existing machine-id
-if [ -f "/etc/redhat-access-insights/machine-id" ]; then
-mv /etc/redhat-access-insights/machine-id /etc/insights-client/machine-id
-fi
-#Migrate existing config
-if [ -f "/etc/redhat-access-insights/redhat-access-insights.conf" ]; then
-mv /etc/redhat-access-insights/redhat-access-insights.conf /etc/insights-client/insights-client.conf
-sed -i 's/\[redhat-access-insights\]/\[insights-client\]/' /etc/insights-client/insights-client.conf
-fi
-#Migrate registration record
-if [ -f "/etc/redhat-access-insights/.registered" ]; then
-mv /etc/redhat-access-insights/.registered /etc/insights-client/.registered
-fi
-#Migrate last upload record
-if [ -f "/etc/redhat-access-insights/.lastupload" ]; then
-mv /etc/redhat-access-insights/.lastupload /etc/insights-client/.lastupload
-fi
-if ! [ -d "/etc/redhat-access-insights" ]; then
-mkdir /etc/redhat-access-insights
-fi
-# Symlink new cron job if the old one exists. Remove the old one
-if [ -f "/etc/cron.daily/redhat-access-insights" ]; then
-rm -f /etc/cron.daily/redhat-access-insights
-ln -sf /etc/insights-client/insights-client.cron /etc/cron.daily/insights-client                               
-fi 
-ln -sf /etc/insights-client/insights-client.conf /etc/redhat-access-insights/redhat-access-insights.conf
-ln -sf /etc/insights-client/insights-client.cron /etc/redhat-access-insights/redhat-access-insights.cron
-ln -sf /etc/insights-client/.registered /etc/redhat-access-insights/.registered
-ln -sf /etc/insights-client/.unregistered /etc/redhat-access-insights/.unregistered
-ln -sf /etc/insights-client/machine-id /etc/redhat-access-insights/machine-id
 
 # if the logging directory isnt created then make it
 # and set the ACLs
